+++ packages/expo/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - [cli] Added modules for interacting with Apple and Android platforms. ([#16516](https://github.com/expo/expo/pull/16516) by [@EvanBacon](https://github.com/EvanBacon))
+- [cli] Added modules for creating dev server URLs, akin to `UrlUtils` in `xdl`.
 
 ### 🐛 Bug fixes